value_wei must be an int
Fixes bug introduced in 5098d1a7b06d0003ca6bdbf9452f9dd183bc275e

+++ src/pyetheroll.py
@@ -363,7 +363,7 @@ class Etheroll:
         # `w3.toWei` one has some issues on Android, see:
         # https://github.com/AndreMiras/EtherollApp/issues/77
         # value_wei = w3.toWei(bet_size_ether, 'ether')
-        value_wei = bet_size_ether * 1e18
+        value_wei = int(bet_size_ether * 1e18)
         gas = 310000
         gas_price = w3.toWei(4, 'gwei')
         # since Account.load is hanging while decrypting the password
